Removes older Data Explorer class
